requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
setuptools==68.2.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.5.0
django-webpack-loader==2.0.1
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0
